cloud_functions/daily_insights/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-pubsub==2.*
google-cloud-aiplatform==1.*
google-generativeai==0.*
google-genai==1.*
sendgrid==6.*
python-dotenv==1.*
pandas==2.*
db-dtypes==1.*
